crawler/requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
selectolax>=0.3
PyMuPDF>=1.24
python-dotenv>=1.0
